refactor: extract AI orchestration and network logic from AiWebParser
Extracted environment-agnostic AI logic (`buildAiPayload`, `extractAiResponse`,
`callAiGenerate`, etc.) from `AiWebParser` to `SharedCore`. Moved environment-
specific HTTP logic (`fetch`, `Request`) to `web-adapter` and `scriptable-adapter`.
Updated parser signatures to accept `httpAdapter` via dependency injection,
enforcing pure parser logic.
