+++ scripts/parsers/chunk-parser.js
@@ -32,7 +32,7 @@ class ChunkParser {
     }
 
     // Main parsing method - receives HTML data and returns events + additional links
-    parseEvents(htmlData = {}, parserConfig = {}, cityConfig = null, pageClassification = null) {
+    parseEvents(htmlData = {}, parserConfig = {}, cityConfig = null, pageClassification = null, httpAdapter = null) {
         try {
             const events = [];
             const html = htmlData.html;
